requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
ccxt==4.4.23
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
config==0.5.1
cryptography==43.0.3
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.0
pycares==4.4.0
pycparser==2.22
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.2.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.16.0
